third_party/blink/web_tests/external/wpt/webrtc/RTCRtpParameters-transactionId-expected.txt

This is a testharness.js-based test.
FAIL sender.getParameters() should return the same transaction ID if called back-to-back without relinquishing the event loop, even if the microtask queue runs assert_true: expected true got false
FAIL sender.getParameters() should return the same transaction ID if called back-to-back without relinquishing the event loop, even if there is an intervening call to setParameters assert_true: expected true got false
PASS sender.getParameters() should return a different transaction ID if the event loop is relinquished between multiple calls
PASS sender.setParameters() with transaction ID different from last getParameters() should reject with InvalidModificationError
PASS sender.setParameters() with transaction ID unset should reject with TypeError
PASS setParameters() twice with the same parameters should reject with InvalidStateError
PASS setParameters() with parameters older than last getParameters() should reject with InvalidModificationError
FAIL setParameters() when the event loop has been relinquished since the last getParameters() should reject with InvalidStateError assert_unreached: Should have rejected: undefined Reached unreachable code
Harness: the test ran to completion.
